Remote/src/test/java/com/example/remote/mapper/ProjectsResponseModelMapperTest.kt
package com.example.remote.mapper

import com.example.remote.factory.TestProjectDataFactory
import org.junit.Test
import kotlin.test.assertEquals

class ProjectsResponseModelMapperTest {

    val mapper = ProjectsResponseModelMapper()

    @Test
    fun mapFromModelMapsData() {
        val model = TestProjectDataFactory.makeProjectModel()
        val entity = mapper.mapFromModel(model)

        assertEquals(model.id, entity.id)
        assertEquals(model.name, entity.name)
        assertEquals(model.fullName, entity.fullName)
        assertEquals(model.starCount.toString(), entity.starCount)
        assertEquals(model.dateCreated, entity.dateCreated)
        assertEquals(model.owner.ownerName, entity.ownerName)
        assertEquals(model.owner.ownerAvatar, entity.ownerAvatar)
    }
}